requirements.txt
bilibili-api-python
pandas
numpy
scipy
scikit-learn
statsmodels
matplotlib
wordcloud
jieba
transformers
torch